fix(Attr): Resolve circular depenedencies
