+++ lib/jsdom/living/attributes.js
@@ -1,6 +1,6 @@
 "use strict";
 const DOMException = require("domexception/webidl2js-wrapper");
-const nodeImpl = require("./nodes/Node-impl.js");
+const { appendAttribute, changeAttribute } = require("./elements.js");
 const attrGenerated = require("./generated/Attr");
 const { asciiLowercase } = require("./helpers/strings");
 const { HTML_NS } = require("./helpers/namespaces");
@@ -23,31 +23,6 @@ exports.hasAttributeByNameNS = function (element, namespace, localName) {
   return element._attributeList.some(attribute => {
     return attribute._localName === localName && attribute._namespace === namespace;
   });
-};
-
-exports.appendAttribute = function (element, attribute) {
-  // https://dom.spec.whatwg.org/#concept-element-attributes-append
-
-  const { _localName, _namespace } = attribute;
-  queueAttributeMutationRecord(element, _localName, _namespace, null);
-
-  const attributeList = element._attributeList;
-
-  attributeList.push(attribute);
-  attribute._element = element;
-
-  // Sync name cache
-  const name = attribute._qualifiedName;
-  const cache = element._attributesByNameMap;
-  let entry = cache.get(name);
-  if (!entry) {
-    entry = [];
-    cache.set(name, entry);
-  }
-  entry.push(attribute);
-
-  // Run jsdom hooks; roughly correspond to spec's "An attribute is set and an attribute is added."
-  element._attrModified(name, attribute._value, null);
 };
 
 exports.removeAttribute = function (element, attribute) {
@@ -184,7 +159,7 @@ exports.setAttribute = function (element, attr) {
   if (oldAttr !== null) {
     exports.replaceAttribute(element, oldAttr, attr);
   } else {
-    exports.appendAttribute(element, attr);
+    appendAttribute(element, attr);
   }
 
   return oldAttr;
@@ -208,12 +183,12 @@ exports.setAttributeValue = function (element, localName, value, prefix, namespa
       localName,
       value
     });
-    exports.appendAttribute(element, newAttribute);
+    appendAttribute(element, newAttribute);
 
     return;
   }
 
-  nodeImpl.changeAttribute(element, attribute, value);
+  changeAttribute(element, attribute, value);
 };
 
 exports.removeAttributeByName = function (element, name) {
