+++ lib/jsdom/living/node.js
@@ -1,6 +1,5 @@
 "use strict";
-const attributes = require("./attributes");
-const attrGenerated = require("./generated/Attr.js");
+const { appendAttribute } = require("./elements.js");
 const { cloningSteps, domSymbolTree } = require("./helpers/internal-constants");
 const NODE_TYPE = require("./node-type");
 const orderedSetParse = require("./helpers/ordered-set").parse;
@@ -32,7 +31,11 @@ module.exports.clone = function (node, document, cloneChildren) {
     case NODE_TYPE.ELEMENT_NODE:
       copy = document._createElementWithCorrectElementInterface(node._localName, node._namespaceURI);
       copy._prefix = node._prefix;
-      copyAttributeList(node, copy);
+      copyAttributeList(node, copy, document);
+      break;
+
+    case NODE_TYPE.ATTRIBUTE_NODE:
+      copy = document._createAttribute(node._localName, node._namespace, node._namespacePrefix, node._value);
       break;
 
     case NODE_TYPE.TEXT_NODE:
@@ -70,18 +73,18 @@ module.exports.clone = function (node, document, cloneChildren) {
   return copy;
 };
 
-function copyAttributeList(sourceElement, destElement) {
+function copyAttributeList(sourceElement, destElement, document) {
   // Needed by https://dom.spec.whatwg.org/#concept-node-clone
 
   for (const sourceAttr of sourceElement._attributeList) {
-    const destAttr = attrGenerated.createImpl(destElement._globalObject, [], {
-      namespace: sourceAttr._namespace,
-      namespacePrefix: sourceAttr._namespacePrefix,
-      localName: sourceAttr._localName,
-      value: sourceAttr._value
-    });
+    const destAttr = document._createAttribute(
+      sourceAttr._localName,
+      sourceAttr._namespace,
+      sourceAttr._namespacePrefix,
+      sourceAttr._value
+    );
 
-    attributes.appendAttribute(destElement, destAttr);
+    appendAttribute(destElement, destAttr);
   }
 }
