+++ lib/jsdom/living/nodes/Document-impl.js
@@ -747,7 +747,7 @@ class DocumentImpl extends NodeImpl {
       localName = asciiLowercase(localName);
     }
 
-    return generatedAttr.createImpl(this._globalObject, [], { localName });
+    return this._createAttribute(localName);
   }
 
   createAttributeNS(namespace, name) {
@@ -757,10 +757,15 @@ class DocumentImpl extends NodeImpl {
     namespace = namespace !== null ? String(namespace) : namespace;
 
     const extracted = validateAndExtract(this._globalObject, namespace, name);
+    return this._createAttribute(extracted.localName, extracted.namespace, extracted.prefix);
+  }
+
+  _createAttribute(localName, namespace, namespacePrefix, value) {
     return generatedAttr.createImpl(this._globalObject, [], {
-      namespace: extracted.namespace,
-      namespacePrefix: extracted.prefix,
-      localName: extracted.localName
+      namespace,
+      namespacePrefix,
+      localName,
+      value
     });
   }
