+++ lib/jsdom/living/nodes/Element-impl.js
@@ -3,9 +3,10 @@ const { addNwsapi } = require("../helpers/selectors");
 const { HTML_NS } = require("../helpers/namespaces");
 const { mixin, memoizeQuery } = require("../../utils");
 const idlUtils = require("../generated/utils");
-const { changeAttribute, implementation: NodeImpl } = require("./Node-impl");
+const NodeImpl = require("./Node-impl").implementation;
 const ParentNodeImpl = require("./ParentNode-impl").implementation;
 const ChildNodeImpl = require("./ChildNode-impl").implementation;
+const { appendAttribute, changeAttribute } = require("../elements.js");
 const attributes = require("../attributes");
 const namedPropertiesWindow = require("../named-properties-window");
 const NODE_TYPE = require("../node-type");
@@ -241,7 +242,7 @@ class ElementImpl extends NodeImpl {
 
     if (attribute === null) {
       const newAttr = attrGenerated.createImpl(this._globalObject, [], { localName: name, value });
-      attributes.appendAttribute(this, newAttr);
+      appendAttribute(this, newAttr);
       return;
     }
 
@@ -278,7 +279,7 @@ class ElementImpl extends NodeImpl {
     if (attribute === null) {
       if (force === undefined || force === true) {
         const newAttr = attrGenerated.createImpl(this._globalObject, [], { localName: qualifiedName, value: "" });
-        attributes.appendAttribute(this, newAttr);
+        appendAttribute(this, newAttr);
         return true;
       }
       return false;
