+++ lib/jsdom/living/nodes/Node-impl.js
@@ -7,7 +7,7 @@ const { simultaneousIterators } = require("../../utils");
 const NODE_TYPE = require("../node-type");
 const NODE_DOCUMENT_POSITION = require("../node-document-position");
 const { clone, locateNamespacePrefix, locateNamespace } = require("../node");
-const { queueAttributeMutationRecord } = require("../helpers/mutation-observers");
+const { changeAttribute } = require("../elements.js");
 
 const NodeList = require("../generated/NodeList");
 
@@ -1140,18 +1140,6 @@ exports.setAnExistingAttributeValue = (attribute, value) => {
   if (element === null) {
     attribute._value = value;
   } else {
-    exports.changeAttribute(element, attribute, value);
+    changeAttribute(element, attribute, value);
   }
-};
-
-// https://dom.spec.whatwg.org/#concept-element-attributes-change
-exports.changeAttribute = (element, attribute, value) => {
-  const { _localName, _namespace, _value } = attribute;
-  queueAttributeMutationRecord(element, _localName, _namespace, _value);
-
-  const oldValue = attribute._value;
-  attribute._value = value;
-
-  // Run jsdom hooks; roughly correspond to spec's "An attribute is set and an attribute is changed."
-  element._attrModified(attribute._qualifiedName, value, oldValue);
 };
